Changes to DDL 'Role' table & Added Procedures
Problem found in updating the primary key of the role table which I've
changed to RoleType (as attributes apparently can't have the same name
as their table, and for consistency: ie AssessmentType and MeetingType
tables). Will explain more tomorrow.
Also more procedures added. And I think we need to change the 'view'
procedures into functions that return the values in the table, since the
dbms_putline lines are only for helping us in testing.