NEW FORMS
Created a basic Employee Dashboard, with basic menu. Menu needs to be
implemented and the remaining forms created. Also we need to figure out
what sql functs/procs we actually need to write, by understanding what
parts the front end will cover, ie how it will formulate sql commands
etc. We need to implement the hard coded things discussed in the last
meeting with Ashir (the column names and types etc). Once we get it
working we can use it as a template to create the rest of the functions
for other forms (which we can also create by using some of the templates
for the ones we've started, and can refer to the diagrams in the fb
chat) more efficiently. We're starting to get somewhere guys :+1: