Unit Offerings Now Work, SQL changes
SQL: So we need to work out what we do with DDL, eg if we keep the
'EmployeeRole' table since the information it 'would' contain is
actually stored in the Unit Offering (as it has a convenor) and Team (a
supervisor) tables as we discussed that time - Sam's point being that it
doesn't matter where it's stored, we can use a join if we need to find
all roles. It's practical to assign a role to a specific moment in time
in this case.

But yes, we then need to change the functions/procedures once we've
finalised the DDL. A whole heap of changes, since the primary keys and
paramaters don't all match up and everything (they run, they just won't
function as we want them to)  :100:  :+1:

THEN MORE FORMS. :8ball: