Project Working and other changes
Projects can be added, edits and deleted.

Multiple changes to Enrolment forms, still working on it. Mainly on the
combo boxes to get them to load their data intuitively.

In UnitOfferings, Employees are now loaded into the combo box from the
EmployeeRole table, only if they are convenors, fixing the errors that
occurred if you chose an employee who was not a registered convenor.