COMBO BOXES WORK.
Pain in the butt but they are working properly in Enrolments, and now
need to be implemented for the other forms (namely project, assessment,
team, meeting...), which thankfully should be a matter of copying the
code (as the objects names on the form are all the same).

We can use the same concepts for other combo boxes too, for eg with
meetings you could pick the team in a similar way, picking one combo box
at once.