DDL and Procedure Changes
Tables AssessmentTask and AssessmentCriterion both now have date columns
called DueDate, as Tasks (for team contribution) and Criteria (for peer
assessment) are extensions of the single master version of assessment
that the convenor creates.

Tables StudentHours and StudentRatings now have columns called
DateSubmitted, as these will exist for each student (in fact multiple
for each student) and are what the student submits their scores/hours
to.