Trying to restore...
Cos github is having major issues...

Problem seems to be with the frmTeamContribution as the form won't
display, and the designer is full of "head <<<<<" blocks. Even when they
are removed though it has a ton of errors where it can't recognise the
objects that the designer refers to... So I've saved a copy of the
former code (both cs file and the designer file, with all the
additions), but will reload the old stable version of the form.

Have no idea wth is going on with the suo and cache files though...