Login is implemented!
You can now login as any one of the user accounts that exist in the
database, and the access level is retrieved from there too.

Currently it's only really noticeable with the student and employee
details, where you can only view 'all' the details if you are an admin,
otherwise you're only able to see your own.