Big changes
So bar a few of the forms that are still being developed, nearly all of
the forms open differently/specifically for users, depending on their
role. This also affects the data that loads.

Changes to SQL: view functions for unitoffering, project and team so
far, have been updated so that you pass two parameters, user (which is
either an empid or stuid) and role (which is admin, convenor,
supervisor. student). It then loads the relevant data for that table,
depending on the role, since the information about different roles is
stored all different locations and needs to be matches in slightly
different ways. For example when you want to load teams, you want
students to see any team they're part (studentteamallocation table).
Allocated supervisors are stored in the team table itself, and convenors
are only stored in the unitofferings table, so they should load teams
only for the units they convene. Hopefully that makes sense. But the
point is it works, and is like 95% stable. The current problems at the
moment are dealing with users with multiple roles...